requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==7.4.2
httpx==0.25.0
pytest-asyncio==0.21.1